+++ router/pump.go
@@ -33,6 +33,10 @@ func getopt(name, dfault string) string {
 	return value
 }
 
+func originalTimestamps() bool {
+	return getopt("ORIGINAL_TIMESTAMPS", "false") == "true"
+}
+
 func debug(v ...interface{}) {
 	if os.Getenv("DEBUG") != "" {
 		log.Println(v...)
@@ -244,6 +248,7 @@ func (p *LogsPump) pumpLogs(event *docker.APIEvents, backlog bool, inactivityTim
 				Since:             sinceTime.Unix(),
 				InactivityTimeout: inactivityTimeout,
 				RawTerminal:       rawTerminal,
+				Timestamps:        originalTimestamps(),
 			})
 			if err != nil {
 				debug("pump.pumpLogs():", id, "stopped with error:", err)
@@ -370,10 +375,11 @@ func newContainerPump(container *docker.Container, stdout, stderr io.Reader) *co
 				}
 				return
 			}
+			logMessage, logTime := parseLogLine(line, originalTimestamps())
 			cp.send(&Message{
-				Data:      strings.TrimSuffix(line, "\n"),
+				Data:      logMessage,
 				Container: container,
-				Time:      time.Now(),
+				Time:      logTime,
 				Source:    source,
 			})
 		}
@@ -404,4 +410,23 @@ func (cp *containerPump) remove(logstream chan *Message) {
 	cp.Lock()
 	defer cp.Unlock()
 	delete(cp.logstreams, logstream)
+}
+
+func parseLogLine(line string, originalTimestamps bool) (string, time.Time) {
+	line = strings.TrimSuffix(line, "\n")
+
+	if ! originalTimestamps {
+		return line, time.Now()
+	}
+
+	logEntry := strings.SplitN(line, " ", 2)
+	logTime, err := time.Parse(time.RFC3339Nano, logEntry[0])
+	if err != nil {
+		return line, time.Now()
+	}
+
+	if len(logEntry) == 2 {
+		return logEntry[1], logTime
+	}
+	return "", logTime
 }